Move logout component
