+++ todoqueue_frontend/src/App.js
@@ -13,9 +13,9 @@ import ConfirmRegistration from './components/ConfirmRegistration';
 import FailedRegistration from './components/FailedRegistration';
 import ForgotPassword from './components/forgotPassword';
 import Login from './components/Login';
+import { Logout } from './components/logout';
 
 import { Navigation } from './navigation';
-import { Logout } from './logout';
 import { SignUp } from './signup';
 import { ResetPassword } from './resetPassword';
 import { ManageHouseholds } from './households';
